refactor: only one public path for each item in module
